Adding license file.
I would like to include your completions files in my project.  Presumably, you would like them to be open source, but I don't see a license here.  I am adding the apache license.  This may of course not be how you want to license your code.  But please add a license otherwise it is not possible to use your work in large companies.

Thanks for the completions files.  They are quite useful.